refactor(daemon): move unix socket peer identity helpers into lifecycle
unix_socket_peer_pid moves byte-identical from src/main.rs into
nestweaver_daemon::lifecycle, which owns every other ownership proof
(db_write_lock, the pidfile flock), and is already a client dependency —
the client's pidfile-less daemon adoption needs the same kernel evidence.

A Linux-only sibling, unix_socket_peer_cred, exposes the (pid, uid) pair
from the same SO_PEERCRED UCred so adoption can require a same-uid peer.
No new macOS-only surface: LOCAL_PEERPID stays pid-only and no other
platform variant is added. The main.rs callers (socket cleanup and
daemon_socket_reported_pid) are re-pointed.
